+++ src/sumo/wrapper/_blob_client.py
@@ -1,11 +1,12 @@
 import httpx
 
-from ._request_error import raise_request_error_exception
-
+from .decorators import http_unpack, raise_for_status, http_retry
 
 class BlobClient:
     """Upload blobs to blob store using pre-authorized URLs"""
 
+    @raise_for_status
+    @http_retry
     def upload_blob(self, blob: bytes, url: str):
         """Upload a blob.
 
@@ -16,21 +17,17 @@ class BlobClient:
 
         headers = {
             "Content-Type": "application/octet-stream",
-            "Content-Length": str(len(blob)),
             "x-ms-blob-type": "BlockBlob",
         }
 
-        try:
-            response = httpx.put(url, data=blob, headers=headers)
-        except httpx.ProxyError as err:
-            raise_request_error_exception(503, err)
-
-        if response.is_error:
-            raise_request_error_exception(response.status_code, response.text)
+        response = httpx.put(url, content=blob, headers=headers)
 
         return response
 
+    @raise_for_status
+    @http_retry
     async def upload_blob_async(self, blob: bytes, url: str):
+
         """Upload a blob async.
 
         Parameters:
@@ -40,19 +37,12 @@ class BlobClient:
 
         headers = {
             "Content-Type": "application/octet-stream",
-            "Content-Length": str(len(blob)),
             "x-ms-blob-type": "BlockBlob",
         }
 
-        try:
-            async with httpx.AsyncClient() as client:
-                response = await client.put(
-                    url=url, data=blob, headers=headers
-                )
-        except httpx.ProxyError as err:
-            raise_request_error_exception(503, err)
-
-        if response.is_error:
-            raise_request_error_exception(response.status_code, response.text)
+        async with httpx.AsyncClient() as client:
+            response = await client.put(
+                url=url, content=blob, headers=headers
+            )
 
         return response
